requirements.txt
astronomer-cosmos==1.7.1
kaggle==1.6.17
kagglehub==0.3.4
pyyaml==6.0.2
dbt-core==1.5.2
dbt-bigquery==1.5.2
astro-run-dag==0.2.7
google-auth
google-cloud-core
google-cloud-bigquery-storage